requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
pydantic
pyquery
selenium
torch==1.7.1
transformers
tqdm